Experimental 1.8 support
